Update: update API routing for messaging app and update documentation
This commit sets up the URL routing in  to expose API endpoints for users, conversations, messages using Django REST framewok's DefaultRouter, implementing a key part of Task 3.

Additionally both the root  and the  files have been updated to reflect the new API endpoints and to properly list Messaging app as a project within the repository overview.